Added rest calls for getting, creating and removing itinerary

+++ TravelAgencyREST/src/java/ws/dtu/resources/model/IniteraryDatabase.java
@@ -15,11 +15,22 @@ public final class IniteraryDatabase {
     
     static HashMap<Integer,Itinerary> db = new HashMap<Integer, Itinerary>();
     
+    public static Itinerary get(int itinerary_id) throws NoSuchIdentfierException {
+        if(!db.containsKey(itinerary_id)) {
+            throw new NoSuchIdentfierException();
+        }
+        
+        return db.get(itinerary_id);
+    }
+    
     public static void insert (Itinerary itinerary) {
         db.put(new Integer (itinerary.getID()), itinerary);
     }
     
-    public static void delete (int itinerary_id) {
+    public static void delete (int itinerary_id) throws NoSuchIdentfierException {
+        if(!db.containsKey(itinerary_id)) {
+            throw new NoSuchIdentfierException();
+        }
         db.remove(new Integer(itinerary_id));
     }
